+++ neuro-cli/src/neuro_cli/job.py
@@ -232,7 +232,19 @@ async def port_forward(
 
 @command()
 @argument("job", type=JOB)
-async def logs(root: Root, job: str) -> None:
+@option(
+    "--since",
+    metavar="DATE_OR_TIMEDELTA",
+    help="Only return logs after a specific date (including). "
+    "Use value of format '1d2h3m4s' to specify moment in "
+    "past relatively to current time.",
+)
+@option(
+    "--timestamps",
+    is_flag=True,
+    help="Include timestamps on each line in the log output.",
+)
+async def logs(root: Root, since: str, job: str, timestamps: bool) -> None:
     """
     Print the logs for a job.
     """
@@ -241,7 +253,14 @@ async def logs(root: Root, job: str) -> None:
         client=root.client,
         status=JobStatus.items(),
     )
-    await process_logs(root, id, None, cluster_name=cluster_name)
+    await process_logs(
+        root,
+        id,
+        None,
+        cluster_name=cluster_name,
+        since=_parse_date(since),
+        timestamps=timestamps,
+    )
 
 
 @command()
@@ -772,7 +791,7 @@ async def top(
         await asyncio.gather(create_pollers(), renderer())
 
 
-@command()
+@command(deprecated=True, hidden=True)
 @argument("job", type=JOB)
 @argument("image", type=RemoteImageType())
 async def save(root: Root, job: str, image: RemoteImage) -> None: